Clean up some code files.
