+++ Assets/Jacob/Scripts/Controllers/Player.Functionality.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9cb8472cde774bc48c88103066f114e7
+timeCreated: 1682813493
