+++ Assets/Jacob/Scripts/Editor/PlayerEditor.Properties.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 5d0991b027854f86894e71b0a4896d70
+timeCreated: 1682814448
